+++ src/components/TrendsPage.js
@@ -6,6 +6,7 @@ import ThrowerLeaverChart from './ThrowerLeaverChart'
 import StreaksChart from './StreaksChart'
 import GroupSizeChart from './GroupSizeChart'
 import HeroesChart from './HeroesChart'
+import DayTimeChart from './DayTimeChart'
 import MapChart from './MapChart'
 import Match from '../models/Match'
 import Color from '../models/Color'
@@ -70,6 +71,8 @@ class TrendsPage extends Component {
         <GroupSizeChart season={season} matches={matches} />
         <hr className="mb-4 pt-4" />
         <HeroesChart season={season} matches={matches} />
+        <hr className="mb-4 pt-4" />
+        <DayTimeChart season={season} matches={matches} />
       </div>
     )
   }