+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added an 'Edit' option to each account on the Accounts page to allow renaming them
 - Sort accounts alphabetically on the Accounts page, ignoring capital letters
 - Added a chart showing wins and losses by hero to the Trends page
+- Added a chart to the Trends page showing wins and losses by the general time of day and day of week played
 
 ## Version 0.1.3